update.sh: use TLS when downloading rootFS
This commit modifies the URL for the Alpine root FS to use the HTTPS
scheme. It also adjusts the cURL invocation to ignore cert signing
errors, as the server presents a self-signed cert. Even though we cannot
verify the cert, this way the download bits cannot be modified mid
flight once the TLS session has been established.
This script is still vulnerable in other regards, e.g. if DNS results
are corrupt.